I added code to print out an error if the item is out of stock.
The previous version of the code had a bug in it.
If the method sell() was called and int m_in_stock was <= 0,
then it would still decrement the m_in_stock, causing it to be negative.

My version of the code only allows you to decrement m_in_stock if it is
> 0, if there are still items left. If there are no more items left,
then an error message is printed to the user.